Fix Issue #1: Sample Text is not getting properly updated when the language changes.
The code for making a single phonetic change was relying on there
needing to be a spelling change to update the sample text for a given
word.  This resulted in many words not getting updated if the user had
decided to also change the spelling in line with an existing rule.