Fix Issue #16: Adjust the default form size to fit on a 1024x768 screen.
The width of the LanguageHoningForm was reduced slightly to keep it
under 1024 wide.

Code was added to both the LanguageHoningForm and the SoundMapListEditor
to check and see if they exceed 1024x768.  These will throw (unhandled)
exceptions which should be spotted during testing to allow the
developer to update the sizing.